requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.10.0+cpu

beautifulsoup4==4.10.0
bidict==0.21.4
cachelib==0.4.1
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
dnspython==1.16.0
eventlet==0.30.2
Flask==2.0.2
Flask-Session==0.4.0
Flask-SocketIO==5.1.1
gevent==21.8.0
gevent-websocket==0.10.1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
lxml==4.6.4
lyricsgenius==3.0.1
MarkupSafe==2.0.1
nltk==3.6.5
numpy==1.21.4
pandas==1.3.4
python-dateutil==2.8.2
python-dotenv==0.19.2
python-engineio==4.3.0
python-socketio==5.5.0
pytrends==4.7.3
pytz==2021.3
regex==2021.11.10
requests==2.26.0
simple-websocket==0.4.0
six==1.16.0
soupsieve==2.3.1
spotipy==2.19.0
# torch==1.10.0
# torchtext==0.11.0
tqdm==4.62.3
typing-extensions==4.0.0
urllib3==1.26.7
Werkzeug==2.0.2
wsproto==1.0.0
zope.event==4.5.0
zope.interface==5.4.0
